mission-charlie-eda/satellite/requirements.txt
fastapi
uvicorn
sse-starlette
requests
python-dotenv
